2.3.5 Amélioration userboard

+++ commands/Modération/users.js
@@ -73,7 +73,7 @@ class usersCommand extends Command {
                 }
 
         }
-        message.delete();
+        if (message.channel.type === 'text') message.delete();
     }
 }
